[Windows] solve warning C4267 and C4244 issue for template plugin test cases (#13736)
* [Windows] solve warning C4267 and C4244 issue for template plugin

* Solve or WA C4267 and C4244 caused by system header files(numeric and utility)

+++ src/plugins/template/tests/functional/op_reference/gather_nd.cpp
@@ -15,7 +15,7 @@ struct GatherNDParams {
     GatherNDParams(
         const reference_tests::Tensor& dataTensor, const reference_tests::Tensor& indicesTensor, int64_t batchDims,
         const reference_tests::Tensor& expectedTensor, const std::string& testcaseName = "") :
-        dataTensor(dataTensor), indicesTensor(indicesTensor), batchDims(batchDims),
+        dataTensor(dataTensor), indicesTensor(indicesTensor), batchDims(static_cast<int32_t>(batchDims)),
         expectedTensor(expectedTensor), testcaseName(testcaseName) {}
 
     reference_tests::Tensor dataTensor;
